Alteracao no nav bar e design

+++ ambulancia-back/src/main/java/com/example/ambulancia/models/entities/veiculo/Veiculo.java
@@ -29,7 +29,7 @@ public class Veiculo extends BaseEntity {
     @Column(nullable = false)
     private String placaVeic;
     @Column(nullable = false)
-    private Double quilometragem; 
+    private Integer quilometragem; 
     @Column(nullable = false)
     private String classe;
